rename time_util to timeutil
ST1003: should not use underscores in package names
