+++ packages/chain/statemanager/sm_gpa/sm_gpa_utils/block_cache_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/iotaledger/wasp/packages/state"
 	"github.com/iotaledger/wasp/packages/testutil/testlogger"
 	"github.com/iotaledger/wasp/packages/util"
-	"github.com/iotaledger/wasp/packages/util/time_util"
+	"github.com/iotaledger/wasp/packages/util/timeutil"
 )
 
 func TestBlockCacheSimple(t *testing.T) {
@@ -22,7 +22,7 @@ func TestBlockCacheSimple(t *testing.T) {
 
 	factory := NewBlockFactory(t)
 	blocks := factory.GetBlocks(4, 1)
-	blockCache, err := NewBlockCache(time_util.NewDefaultTimeProvider(), 100, NewEmptyTestBlockWAL(), mockStateManagerMetrics(), log)
+	blockCache, err := NewBlockCache(timeutil.NewDefaultTimeProvider(), 100, NewEmptyTestBlockWAL(), mockStateManagerMetrics(), log)
 	require.NoError(t, err)
 	blockCache.AddBlock(blocks[0])
 	blockCache.AddBlock(blocks[1])
@@ -39,7 +39,7 @@ func TestBlockCacheCleaning(t *testing.T) {
 
 	factory := NewBlockFactory(t)
 	blocks := factory.GetBlocks(6, 2)
-	blockCache, err := NewBlockCache(time_util.NewDefaultTimeProvider(), 100, NewEmptyTestBlockWAL(), mockStateManagerMetrics(), log)
+	blockCache, err := NewBlockCache(timeutil.NewDefaultTimeProvider(), 100, NewEmptyTestBlockWAL(), mockStateManagerMetrics(), log)
 	require.NoError(t, err)
 
 	beforeTime := time.Now()
@@ -86,7 +86,7 @@ func TestBlockCacheSameBlockCleaning(t *testing.T) {
 
 	factory := NewBlockFactory(t)
 	blocks := factory.GetBlocks(3, 2)
-	blockCache, err := NewBlockCache(time_util.NewDefaultTimeProvider(), 100, NewEmptyTestBlockWAL(), mockStateManagerMetrics(), log)
+	blockCache, err := NewBlockCache(timeutil.NewDefaultTimeProvider(), 100, NewEmptyTestBlockWAL(), mockStateManagerMetrics(), log)
 	require.NoError(t, err)
 	blockCache.AddBlock(blocks[0])
 	blockCache.AddBlock(blocks[1])
@@ -119,7 +119,7 @@ func TestBlockCacheWAL(t *testing.T) {
 	factory := NewBlockFactory(t)
 	blocks := factory.GetBlocks(3, 2)
 	wal := NewMockedTestBlockWAL()
-	blockCache, err := NewBlockCache(time_util.NewDefaultTimeProvider(), 100, wal, mockStateManagerMetrics(), log)
+	blockCache, err := NewBlockCache(timeutil.NewDefaultTimeProvider(), 100, wal, mockStateManagerMetrics(), log)
 	require.NoError(t, err)
 	blockCache.AddBlock(blocks[0])
 	blockCache.AddBlock(blocks[1])
@@ -144,7 +144,7 @@ func TestBlockCacheCleanUp(t *testing.T) {
 	blocks := factory.GetBlocks(16, 2)
 	now := time.Now()
 	wal := NewEmptyTestBlockWAL()
-	tp := time_util.NewArtificialTimeProvider(now)
+	tp := timeutil.NewArtificialTimeProvider(now)
 	blockCache, err := NewBlockCache(tp, 5, wal, mockStateManagerMetrics(), log)
 	require.NoError(t, err)
 	setNowAddBlockFun := func(d time.Duration, b state.Block) {
@@ -220,7 +220,7 @@ func TestBlockCacheFull(t *testing.T) {
 	blocks := factory.GetBlocks(5, 2)
 	now := time.Now()
 	wal := NewMockedTestBlockWAL()
-	tp := time_util.NewArtificialTimeProvider(now)
+	tp := timeutil.NewArtificialTimeProvider(now)
 	blockCache, err := NewBlockCache(tp, 100, wal, mockStateManagerMetrics(), log)
 	require.NoError(t, err)
 	blockCache.AddBlock(blocks[0]) // Will be dropped from cache AND WAL
