+++ packages/util/timeutil/interface.go
@@ -1,4 +1,4 @@
-package time_util
+package timeutil
 
 import (
 	"time"
