+++ packages/util/timeutil/time_provider_default.go
@@ -1,6 +1,6 @@
 // Package time_util provides utilities for time-related operations,
 // including implementations of time providers for testing and production use.
-package time_util
+package timeutil
 
 import (
 	"time"
